Remove display block and border-top
B/c mixin gets applied to things that are default block or get their display set (to flex for example) so gonna assume it's not needed here

Same for border top. It was requested for things on white backgrounds. Don't want to globally set a border that might be wrong